@@ -36,29 +36,29 @@ dependencies:
   cupertino_icons: ^1.0.2
 
   #Base
-  get: ^4.6.5
+  get: ^4.6.6
   get_storage: ^2.1.1
-  dio: ^5.3.2
-  cached_network_image: ^3.2.3
-  flutter_screenutil: ^5.8.4
-  google_fonts: ^5.1.0
+  dio: ^5.4.0
+  cached_network_image: ^3.3.1
+  flutter_screenutil: ^5.9.0
+  google_fonts: ^6.1.0
   carousel_slider: ^4.2.1
   smooth_page_indicator: ^1.1.0
-  font_awesome_flutter: ^10.5.0
+  font_awesome_flutter: ^10.6.0
   pin_code_fields: ^8.0.1
-  flutter_svg: ^2.0.7
+  flutter_svg: ^2.0.9
   flutter_rating_bar: ^4.0.1
   flutter_switch: ^0.3.2
   like_button: ^2.0.5
   timelines: ^0.1.0
   intl: ^0.18.1
-  syncfusion_flutter_sliders: ^22.2.9
+  syncfusion_flutter_sliders: ^24.1.45
   shimmer: ^3.0.0
-  firebase_auth: ^4.7.3
-  firebase_core: ^2.15.1
-  google_sign_in: ^6.1.4
-  fluttertoast: ^8.2.2
-  flutter_facebook_auth: ^6.0.1
+  firebase_auth: ^4.16.0
+  firebase_core: ^2.24.2
+  google_sign_in: ^6.2.1
+  fluttertoast: ^8.2.4
+  flutter_facebook_auth: ^6.0.3
 
 dev_dependencies:
   flutter_test:
