@@ -21,7 +21,7 @@ class CategoryController extends GetxController {
     categoryList = AppArray().categoryList;
 
     update();
-    await Future.delayed(Durations.s1);
+    await Future.delayed(DurationsClass.s1);
     appCtrl.isShimmer = false;
     appCtrl.update();
     update();
