@@ -84,7 +84,7 @@ class ProfileController extends GetxController {
       Get.toNamed(routeName.help);
     }
     update();
-    await Future.delayed(Durations.s1);
+    await Future.delayed(DurationsClass.s1);
     appCtrl.isShimmer =false;
     appCtrl.update();
     Get.forceAppUpdate();
