@@ -26,7 +26,7 @@ class MyApp extends StatelessWidget {
       builder: (_, child) => GetMaterialApp(
         builder: (context, widget) {
           return MediaQuery(
-            data: MediaQuery.of(context).copyWith(textScaleFactor: 1.0),
+            data: MediaQuery.of(context).copyWith(textScaler: const TextScaler.linear(1.0)),
             child: widget!,
           );
         },
