@@ -1,6 +1,5 @@
 import 'package:multikart/config.dart';
 
-
 class HomeScreen extends StatefulWidget {
   const HomeScreen({Key? key}) : super(key: key);
 
@@ -15,51 +14,52 @@ class _HomeScreenState extends State<HomeScreen> {
   Widget build(BuildContext context) {
     return GetBuilder<HomeController>(builder: (_) {
       return Directionality(
-        textDirection: homeCtrl.appCtrl.isRTL ||
-            homeCtrl.appCtrl.languageVal == "ar"
-            ? TextDirection.rtl
-            : TextDirection.ltr,
+        textDirection:
+            homeCtrl.appCtrl.isRTL || homeCtrl.appCtrl.languageVal == "ar"
+                ? TextDirection.rtl
+                : TextDirection.ltr,
         child: Scaffold(
             resizeToAvoidBottomInset: false,
-            body:  homeCtrl.appCtrl.isShimmer ? const HomerShimmer() : SingleChildScrollView(
-                child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    mainAxisSize: MainAxisSize.max,
-                    children: const [
-                      //home category list layout
-                      HomeCategoryList(),
-                      // border line layout
+            body: homeCtrl.appCtrl.isShimmer
+                ? const HomerShimmer()
+                : const SingleChildScrollView(
+                    child: Column(
+                        crossAxisAlignment: CrossAxisAlignment.start,
+                        mainAxisSize: MainAxisSize.max,
+                        children: [
+                        //home category list layout
+                        HomeCategoryList(),
+                        // border line layout
 
-                      BorderLineLayout(),
-                      //banner list layout
+                        BorderLineLayout(),
+                        //banner list layout
 
-                      HomeBannerList(),
+                        HomeBannerList(),
 
-                      //deals of the day
-                      HomeDealsOfTheDayLayout(),
+                        //deals of the day
+                        HomeDealsOfTheDayLayout(),
 
-                      // border line layout
-                      BorderLineLayout(),
+                        // border line layout
+                        BorderLineLayout(),
 
-                      //find your style
-                      FindYourStyle(),
+                        //find your style
+                        FindYourStyle(),
 
-                      //offer time banner
-                      OfferTimeLayout(),
+                        //offer time banner
+                        OfferTimeLayout(),
 
-                      //biggest deal of brands
-                      DealsBrands(),
+                        //biggest deal of brands
+                        DealsBrands(),
 
-                      // border line layout
-                      BorderLineLayout(),
+                        // border line layout
+                        BorderLineLayout(),
 
-                      //kids corner
-                      KidsCorner(),
+                        //kids corner
+                        KidsCorner(),
 
-                      //offer corner
-                      OfferCorner()
-                    ]
-                ))),
+                        //offer corner
+                        OfferCorner()
+                      ]))),
       );
     });
   }
