@@ -2,8 +2,9 @@ import '../../../../config.dart';
 
 class DrawerLeadingTitle extends StatelessWidget {
   final dynamic data;
-final int? index;
-  const DrawerLeadingTitle({Key? key, this.data,this.index}) : super(key: key);
+  final int? index;
+
+  const DrawerLeadingTitle({Key? key, this.data, this.index}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -18,8 +19,13 @@ final int? index;
                 children: <Widget>[
                   data['icon'] == "assets/svg/flags.svg"
                       ? SvgPicture.asset(svgAssets.flags, fit: BoxFit.fill)
-                      :data['icon'] == "assets/svg/currency.svg" ?const Icon(Icons.money) :  SvgPicture.asset(data['icon'],
-                          color: appCtrl.appTheme.blackColor)
+                      : data['icon'] == "assets/svg/currency.svg"
+                          ? const Icon(Icons.money)
+                          : SvgPicture.asset(
+                              data['icon'],
+                              colorFilter: ColorFilter.mode(
+                                  appCtrl.appTheme.blackColor, BlendMode.srcIn),
+                            )
                 ]),
             const Space(20, 0),
             Column(
@@ -29,16 +35,19 @@ final int? index;
                     text: data['title'],
                     fontSize: FontSizes.f12,
                     fontWeight: FontWeight.w600),
-                if(index != 0 && index !=1)
-                LatoFontStyle(
-                    text: data['subTitle'],
-                    fontSize: FontSizes.f12,
-                    fontWeight: FontWeight.normal)
+                if (index != 0 && index != 1)
+                  LatoFontStyle(
+                      text: data['subTitle'],
+                      fontSize: FontSizes.f12,
+                      fontWeight: FontWeight.normal)
               ],
             )
           ],
         ),
-        if (data['title'] == "Mode" ||data['title'] == "الوضع" ||data['title'] == "तरीका" ||data['title'] == "방법" )
+        if (data['title'] == "Mode" ||
+            data['title'] == "الوضع" ||
+            data['title'] == "तरीका" ||
+            data['title'] == "방법")
           ThemeSwitcher(
               onToggle: (val) {
                 appCtrl.isTheme = val;
@@ -46,7 +55,10 @@ final int? index;
                 ThemeService().switchTheme(val);
               },
               status2: appCtrl.isTheme),
-        if (data['title'] == "RTL" ||data['title'] == "아르 자형티엘"  || data['title'] == "आरटीएल"  ||data['title'] == "رتيإل" )
+        if (data['title'] == "RTL" ||
+            data['title'] == "아르 자형티엘" ||
+            data['title'] == "आरटीएल" ||
+            data['title'] == "رتيإل")
           ThemeSwitcher(
               onToggle: (val) {
                 appCtrl.isRTL = val;
