@@ -6,8 +6,8 @@ class FormTextBox extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Column(
-      children:const [
+    return const Column(
+      children:[
         CountryTextBox(),
         Space(0, 30),
         FullNameTextBox(),
