@@ -40,7 +40,8 @@ class CurrencyBottomSheet extends StatelessWidget {
                           e['icon'].toString(),
                           height: AppScreenUtil().screenHeight(25),
                           fit: BoxFit.contain,
-                          color: appCtrl.appTheme.blackColor,
+                          colorFilter: ColorFilter.mode(
+                              appCtrl.appTheme.blackColor, BlendMode.srcIn),
                         ),
                         const Space(10, 0),
                         LatoFontStyle(
