@@ -8,7 +8,8 @@ class DeleteIcon extends StatelessWidget {
     return SvgPicture.asset(
       svgAssets.delete,
       height: AppScreenUtil().screenHeight(18),
-      color: appCtrl.isTheme ? appCtrl.appTheme.blackColor : appCtrl.appTheme.whiteColor,
+      colorFilter: ColorFilter.mode(
+          appCtrl.isTheme ? appCtrl.appTheme.blackColor : appCtrl.appTheme.whiteColor, BlendMode.srcIn),
     );
   }
 }
