@@ -17,7 +17,7 @@ class LatoFontStyle extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return InkWell(
+    return GestureDetector(
         onTap: onTap,
         child: Text(text!,
             overflow: overflow,
